Fix leak on rejected UDP destinations
Currently, if a UDP packet decrypts successfully, but is rejected due to
the contents of the address field (e.g. destination is localhost), the
NAT will add an entry for this packet (including a UDP connection and a
goroutine), but will not set a deadline to expire the mapping.  This
results in a leak.

In this change, every connection added to the mapping has a default
deadline, so that nothing lingers forever in the map.