Refactor (tcpService).handleConnection to avoid FD leaks.
We (Lantern) found large amounts of (dead) socket FD's open on the shadowsocks
free tracks and after doing a coredump, we could see a lot of connections
stuck inside net.Pipe and, but as a bigger smoking gun, a lot inside
the `connError := func() *onet.ConnectionError {` function.

It seems that if the connect closes only in one direction and hangs
in the other, the connection will never close, even if the underlying
connection socket dies and times out. This is because a deadlock forms
inside the function waiting for the other direction to die.

To prevent this, we simply just make both go functions and see what one
errors first with a select{}.